tests/wpt/tests/css/css-transforms/transform_translate_neg.html

<!DOCTYPE html>
<html>
<head>
    <title>CSS Transforms API Test: transform translate with negative translation value</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#transform-property">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions">
    <meta name="flags" content="dom">
    <meta name="assert" content="Check if transform supports translate(-1px, -1px)">
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
    test(function() {
        document.getElementById("test").style.transform = "translate(-1px, -1px)";
        var value = window.getComputedStyle(document.getElementById("test")).getPropertyValue("transform");
        assert_equals(value, "matrix(1, 0, 0, 1, -1, -1)")
    }, "transform_translate_-1px_-1px");
    </script>
</body>
</html>
